refactoring and completing a bit the most simple node store
also renamed it to boxed_components to be more expressive
