+++ hyper_ast/src/store/nodes/mod.rs
@@ -3,19 +3,19 @@ pub mod fetched;
 pub mod hecs;
 #[cfg(feature = "legion")]
 pub mod legion;
-mod simple;
+pub mod boxed_components;
 
 #[cfg(feature = "legion")]
 pub type DefaultNodeStore = legion::NodeStore;
 #[cfg(not(feature = "legion"))]
-pub type DefaultNodeStore = simple::NodeStore;
+pub type DefaultNodeStore = boxed_components::NodeStore;
 
 #[cfg(feature = "legion")]
 pub type DefaultNodeIdentifier = legion::NodeIdentifier;
 #[cfg(not(feature = "legion"))]
-pub type DefaultNodeIdentifier = simple::NodeIdentifier;
+pub type DefaultNodeIdentifier = boxed_components::NodeIdentifier;
 
 #[cfg(feature = "legion")]
 pub type HashedNodeRef<'store> = legion::HashedNodeRef<'store, DefaultNodeIdentifier>;
 #[cfg(not(feature = "legion"))]
-pub type HashedNodeRef<'store> = simple::HashedNodeRef<'store, DefaultNodeIdentifier>;
+pub type HashedNodeRef<'store> = boxed_components::HashedNodeRef<'store, DefaultNodeIdentifier>;
